Update adapter tests to work. Move adapter utilities to dbt.test.util.
