--- core/dbt/tests/__init__.py
@@ -1,1 +1,0 @@
-# dbt.tests directory
