+++ editable-requirements.txt
@@ -1,2 +1,3 @@
 -e ./core
 -e ./plugins/postgres
+-e ./tests/adapter
