+++ tox.ini
@@ -18,6 +18,7 @@ passenv = DBT_* POSTGRES_TEST_* PYTEST_ADDOPTS
 commands =
   {envpython} -m pytest --cov=core -m profile_postgres {posargs} test/integration
   {envpython} -m pytest --cov=core {posargs} tests/functional
+  {envpython} -m pytest --cov=core {posargs} tests/adapter
 
 deps =
   -rdev-requirements.txt
